discovery: Add initial TCP/IP support
The current implementation uses 'limited broadcasts' for device
discovery. The disadvantages are that they are not routed and that
device discovery may not work reliable if multiple network interfaces
are available. Furthermore, they do not work on FreeBSD.

If required, support for device discovery with 'directed broadcasts'
will be added in the future.

Signed-off-by: Marc Schink <jaylink-dev@marcschink.de>