Work around serial number issues.
The firmware "J-Link ARM-OB SAM7 compiled Jun  1 2012 09:33:42",
which is used in the Renesas Promotion Board RPBRX62N for example,
sometimes times out on reading the serial number which makes the device
undiscoverable although it works properly for the rest.

It appears that SEGGER handles this by just ignoring the timeout.
This patch implements the same behaviour. Additionally, the API of
jaylink_device_get_serial_number() is changed to indicate when the
serial number of a device is not available.

Signed-off-by: Andreas Färber <afaerber@suse.de>
Tested-by: Andreas Färber <afaerber@suse.de>
[Add valid_serial_number field and change API.]
Signed-off-by: Marc Schink <jaylink-dev@marcschink.de>