Simplify jaylink_(un)register() API
The additional information provided by jaylink_(un)register() is not yet
documented and not at all necessary to (un)register a connection. Also,
the API in its current state is not well thought out.

Remove access to the additional information in favour of a simpler API.

If necessary, the access may be re-added with a proper documentation and
well thought out API in a future version.

Signed-off-by: Marc Schink <jaylink-dev@marcschink.de>